+++ chain/cosmos/chain_node.go
@@ -25,10 +25,8 @@ import (
 	libclient "github.com/cometbft/cometbft/rpc/jsonrpc/client"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
-	"github.com/cosmos/cosmos-sdk/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authTx "github.com/cosmos/cosmos-sdk/x/auth/tx"
-	paramsutils "github.com/cosmos/cosmos-sdk/x/params/client/utils"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	volumetypes "github.com/docker/docker/api/types/volume"
 	dockerclient "github.com/docker/docker/client"
@@ -676,7 +674,7 @@ func (tn *ChainNode) IsAboveSDK47(ctx context.Context) bool {
 }
 
 // AddGenesisAccount adds a genesis account for each key
-func (tn *ChainNode) AddGenesisAccount(ctx context.Context, address string, genesisAmount []types.Coin) error {
+func (tn *ChainNode) AddGenesisAccount(ctx context.Context, address string, genesisAmount []sdk.Coin) error {
 	amount := ""
 	for i, coin := range genesisAmount {
 		if i != 0 {
@@ -710,7 +708,7 @@ func (tn *ChainNode) AddGenesisAccount(ctx context.Context, address string, gene
 }
 
 // Gentx generates the gentx for a given node
-func (tn *ChainNode) Gentx(ctx context.Context, name string, genesisSelfDelegation types.Coin) error {
+func (tn *ChainNode) Gentx(ctx context.Context, name string, genesisSelfDelegation sdk.Coin) error {
 	tn.lock.Lock()
 	defer tn.lock.Unlock()
 
@@ -834,9 +832,9 @@ func (tn *ChainNode) StoreContract(ctx context.Context, keyName string, fileName
 	return res.CodeInfos[0].CodeID, nil
 }
 
-func (tn *ChainNode) GetTransaction(clientCtx client.Context, txHash string) (*types.TxResponse, error) {
+func (tn *ChainNode) GetTransaction(clientCtx client.Context, txHash string) (*sdk.TxResponse, error) {
 	// Retry because sometimes the tx is not committed to state yet.
-	var txResp *types.TxResponse
+	var txResp *sdk.TxResponse
 	err := retry.Do(func() error {
 		var err error
 		txResp, err = authTx.QueryTx(clientCtx, txHash)
@@ -981,18 +979,18 @@ func (tn *ChainNode) InstantiateContract(ctx context.Context, keyName string, co
 }
 
 // ExecuteContract executes a contract transaction with a message using it's address.
-func (tn *ChainNode) ExecuteContract(ctx context.Context, keyName string, contractAddress string, message string, extraExecTxArgs ...string) (res *types.TxResponse, err error) {
+func (tn *ChainNode) ExecuteContract(ctx context.Context, keyName string, contractAddress string, message string, extraExecTxArgs ...string) (res *sdk.TxResponse, err error) {
 	cmd := []string{"wasm", "execute", contractAddress, message}
 	cmd = append(cmd, extraExecTxArgs...)
 
 	txHash, err := tn.ExecTx(ctx, keyName, cmd...)
 	if err != nil {
-		return &types.TxResponse{}, err
+		return &sdk.TxResponse{}, err
 	}
 
 	txResp, err := tn.GetTransaction(tn.CliContext(), txHash)
 	if err != nil {
-		return &types.TxResponse{}, fmt.Errorf("failed to get transaction %s: %w", txHash, err)
+		return &sdk.TxResponse{}, fmt.Errorf("failed to get transaction %s: %w", txHash, err)
 	}
 
 	if txResp.Code != 0 {
@@ -1091,122 +1089,6 @@ func (tn *ChainNode) GetModuleAccount(ctx context.Context, moduleName string) (Q
 		return QueryModuleAccountResponse{}, err
 	}
 	return queryRes, nil
-}
-
-// VoteOnProposal submits a vote for the specified proposal.
-func (tn *ChainNode) VoteOnProposal(ctx context.Context, keyName string, proposalID string, vote string) error {
-	_, err := tn.ExecTx(ctx, keyName,
-		"gov", "vote",
-		proposalID, vote, "--gas", "auto",
-	)
-	return err
-}
-
-// QueryProposal returns the state and details of a governance proposal.
-func (tn *ChainNode) QueryProposal(ctx context.Context, proposalID string) (*ProposalResponse, error) {
-	stdout, _, err := tn.ExecQuery(ctx, "gov", "proposal", proposalID)
-	if err != nil {
-		return nil, err
-	}
-	var proposal ProposalResponse
-	err = json.Unmarshal(stdout, &proposal)
-	if err != nil {
-		return nil, err
-	}
-	return &proposal, nil
-}
-
-// QueryProposal returns the state and details of an IBC-Go v8 / SDK v50 governance proposal.
-func (tn *ChainNode) QueryProposalV8(ctx context.Context, proposalID string) (*ProposalResponseV8, error) {
-	stdout, _, err := tn.ExecQuery(ctx, "gov", "proposal", proposalID)
-	if err != nil {
-		return nil, err
-	}
-	var proposal ProposalResponseV8
-	err = json.Unmarshal(stdout, &proposal)
-	if err != nil {
-		return nil, err
-	}
-	return &proposal, nil
-}
-
-// SubmitProposal submits a gov v1 proposal to the chain.
-func (tn *ChainNode) SubmitProposal(ctx context.Context, keyName string, prop TxProposalv1) (string, error) {
-	// Write msg to container
-	file := "proposal.json"
-	propJson, err := json.MarshalIndent(prop, "", " ")
-	if err != nil {
-		return "", err
-	}
-	fw := dockerutil.NewFileWriter(tn.logger(), tn.DockerClient, tn.TestName)
-	if err := fw.WriteFile(ctx, tn.VolumeName, file, propJson); err != nil {
-		return "", fmt.Errorf("writing contract file to docker volume: %w", err)
-	}
-
-	command := []string{
-		"gov", "submit-proposal",
-		path.Join(tn.HomeDir(), file), "--gas", "auto",
-	}
-
-	return tn.ExecTx(ctx, keyName, command...)
-}
-
-// UpgradeProposal submits a software-upgrade governance proposal to the chain.
-func (tn *ChainNode) UpgradeProposal(ctx context.Context, keyName string, prop SoftwareUpgradeProposal) (string, error) {
-	command := []string{
-		"gov", "submit-proposal",
-		"software-upgrade", prop.Name,
-		"--upgrade-height", strconv.FormatUint(prop.Height, 10),
-		"--title", prop.Title,
-		"--description", prop.Description,
-		"--deposit", prop.Deposit,
-	}
-
-	if prop.Info != "" {
-		command = append(command, "--upgrade-info", prop.Info)
-	}
-
-	return tn.ExecTx(ctx, keyName, command...)
-}
-
-// TextProposal submits a text governance proposal to the chain.
-func (tn *ChainNode) TextProposal(ctx context.Context, keyName string, prop TextProposal) (string, error) {
-	command := []string{
-		"gov", "submit-proposal",
-		"--type", "text",
-		"--title", prop.Title,
-		"--description", prop.Description,
-		"--deposit", prop.Deposit,
-	}
-	if prop.Expedited {
-		command = append(command, "--is-expedited=true")
-	}
-	return tn.ExecTx(ctx, keyName, command...)
-}
-
-// ParamChangeProposal submits a param change proposal to the chain, signed by keyName.
-func (tn *ChainNode) ParamChangeProposal(ctx context.Context, keyName string, prop *paramsutils.ParamChangeProposalJSON) (string, error) {
-	content, err := json.Marshal(prop)
-	if err != nil {
-		return "", err
-	}
-
-	hash := sha256.Sum256(content)
-	proposalFilename := fmt.Sprintf("%x.json", hash)
-	err = tn.WriteFile(ctx, content, proposalFilename)
-	if err != nil {
-		return "", fmt.Errorf("writing param change proposal: %w", err)
-	}
-
-	proposalPath := filepath.Join(tn.HomeDir(), proposalFilename)
-
-	command := []string{
-		"gov", "submit-proposal",
-		"param-change",
-		proposalPath,
-	}
-
-	return tn.ExecTx(ctx, keyName, command...)
 }
 
 // QueryParam returns the state and details of a subspace param.
@@ -1400,8 +1282,8 @@ func (tn *ChainNode) RemoveContainer(ctx context.Context) error {
 func (tn *ChainNode) InitValidatorGenTx(
 	ctx context.Context,
 	chainType *ibc.ChainConfig,
-	genesisAmounts []types.Coin,
-	genesisSelfDelegation types.Coin,
+	genesisAmounts []sdk.Coin,
+	genesisSelfDelegation sdk.Coin,
 ) error {
 	if err := tn.CreateKey(ctx, valKey); err != nil {
 		return err
